Move extension-instance to its own file
