+++ packages/app/src/cli/models/app/loader.ts
@@ -1,8 +1,9 @@
 import {AppConfigurationSchema, Web, WebConfigurationSchema, App, AppInterface, WebType} from './app.js'
 import {configurationFileNames, dotEnvFileNames} from '../../constants.js'
 import metadata from '../../metadata.js'
-import {ExtensionInstance, ExtensionSpecification} from '../extensions/specification.js'
+import {ExtensionInstance} from '../extensions/extension-instance.js'
 import {TypeSchema} from '../extensions/schemas.js'
+import {ExtensionSpecification} from '../extensions/specification.js'
 import {zod} from '@shopify/cli-kit/node/schema'
 import {fileExists, readFile, glob, findPathUp} from '@shopify/cli-kit/node/fs'
 import {readAndParseDotEnv, DotEnvFile} from '@shopify/cli-kit/node/dot-env'
