+++ packages/app/src/cli/services/dev.ts
@@ -29,7 +29,8 @@ import {getAnalyticsTunnelType} from '../utilities/analytics.js'
 import {buildAppURLForWeb} from '../utilities/app/app-url.js'
 import {HostThemeManager} from '../utilities/host-theme-manager.js'
 
-import {ExtensionInstance, ExtensionSpecification} from '../models/extensions/specification.js'
+import {ExtensionInstance} from '../models/extensions/extension-instance.js'
+import {ExtensionSpecification} from '../models/extensions/specification.js'
 import {Config} from '@oclif/core'
 import {reportAnalyticsEvent} from '@shopify/cli-kit/node/analytics'
 import {execCLI2} from '@shopify/cli-kit/node/ruby'
