+++ packages/app/src/cli/services/function/common.test.ts
@@ -2,7 +2,7 @@ import {inFunctionContext} from './common.js'
 import {load as loadApp} from '../../models/app/loader.js'
 import {testApp, testFunctionExtension} from '../../models/app/app.test-data.js'
 import {AppInterface} from '../../models/app/app.js'
-import {ExtensionInstance} from '../../models/extensions/specification.js'
+import {ExtensionInstance} from '../../models/extensions/extension-instance.js'
 import {describe, vi, expect, beforeEach, test} from 'vitest'
 import {Config} from '@oclif/core'
 import {renderFatalError} from '@shopify/cli-kit/node/ui'
